feat: added create (and its form in create page) and store functions; in Comic model added fillable for store function; changed comics representation from cards to table

+++ database/migrations/2024_03_18_134959_create_comics_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title', 100);
             $table->text('description');
             $table->string('thumb');
-            $table->float('price', 5, 2)->unsigned();
+            $table->decimal('price', 5, 2)->unsigned();
             $table->string('series', 100);
             $table->date('sale_date');
             $table->string('type');
